requirements.txt
mkdocs
mkdocs-git-revision-date-localized-plugin
mkdocs-material
PyYAML
